pyproject.toml
[tool.poetry]
name = "sprite-generator"
version = "0.1.0"
description = "Super simple AI sprite generator"
authors = ["John Loverich <john@glodonusa.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "sprite_generator"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
torch = "^2.0.0"

pillow = "^10.0.0"
click = "^8.1.7"
rembg = "^2.0.50"
numpy = "^1.24.0"
safetensors = "^0.4.0"
diffusers = "^0.33.1"
transformers = "^4.51.3"
accelerate = "^1.6.0"
huggingface-hub = "^0.31.2"
onnxruntime = "^1.22.0"
sentencepiece = "^0.2.0"
streamlit = "^1.45.1"

[tool.poetry.scripts]
sprite-gen = "sprite_generator.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
